+++ src/main/java/baseball/game/stage/AbstractStage.java
@@ -6,7 +6,6 @@ import baseball.domain.object.Referee;
 import baseball.enums.GameState;
 import baseball.game.GameController;
 import baseball.game.domain.GameConfig;
-import baseball.game.domain.ObserveRequest;
 import baseball.view.View;
 
 import java.util.Observable;
