+++ src/main/java/baseball/game/stage/impl/SetUpPitchStage.java
@@ -4,7 +4,7 @@ import baseball.enums.GameState;
 import baseball.game.GameController;
 import baseball.game.stage.AbstractStage;
 import baseball.game.domain.GameConfig;
-import baseball.game.domain.ObserveRequest;
+import baseball.game.stage.ObserveRequest;
 
 public class SetUpPitchStage extends AbstractStage {
     public SetUpPitchStage(GameConfig config) {
